refactor(search): update import paths and simplify search parameters
Update test file import paths to reflect new directory structure
Remove unused activities parameter from intelligent search methods
